Cleanup

+++ src/solo/SoloMesh.h
@@ -30,36 +30,21 @@ namespace solo {
         virtual auto addDynamicVertexBuffer(const VertexBufferLayout &layout, const vec<float> &data, u32 vertexCount) -> u32;
         virtual void updateVertexBuffer(u32 index, u32 vertexOffset, const void *data, u32 vertexCount) {}
         virtual void removeVertexBuffer(u32 index);
-        auto vertexBufferCount() const -> u32 {
-            return layouts_.size();
-        }
-        auto vertexBufferVertexCount(u32 index) const -> u32 {
-            return vertexCounts_.at(index);
-        }
-        auto vertexBufferLayout(u32 index) const -> VertexBufferLayout {
-            return layouts_.at(index);
-        }
+
+        auto vertexBufferCount() const -> u32 { return static_cast<u32>(layouts_.size()); }
+        auto vertexBufferVertexCount(u32 index) const -> u32 { return vertexCounts_.at(index); }
+        auto vertexBufferLayout(u32 index) const -> VertexBufferLayout { return layouts_.at(index); }
         auto vertexBufferData(u32 index) const -> const vec<float> & { return vertexData_.at(index); }
 
         virtual auto addIndexBuffer(const vec<u32> &data, u32 elementCount) -> u32;
         virtual void removeIndexBuffer(u32 index);
-        auto indexBufferCount() const -> u32 {
-            return static_cast<u32>(indexElementCounts_.size());
-        }
-        auto indexBufferElementCount(u32 index) const -> u32 {
-            return indexElementCounts_.at(index);
-        }
-        auto indexBufferElementSize(u32 index) const -> IndexElementSize {
-            return IndexElementSize::Bits32;    // TODO 16-bit support?
-        }
+        auto indexBufferCount() const -> u32 { return static_cast<u32>(indexElementCounts_.size()); }
+        auto indexBufferElementCount(u32 index) const -> u32 { return indexElementCounts_.at(index); }
+        auto indexBufferElementSize(u32 index) const -> IndexElementSize { return IndexElementSize::Bits32; } // TODO 16-bit support?
         auto indexData(u32 index) const -> const vec<u32> & { return indexData_.at(index); }
 
-        auto primitiveType() const -> PrimitiveType {
-            return primitiveType_;
-        }
-        void setPrimitiveType(PrimitiveType type) {
-            primitiveType_ = type;
-        }
+        auto primitiveType() const -> PrimitiveType { return primitiveType_; }
+        void setPrimitiveType(PrimitiveType type) { primitiveType_ = type; }
 
     protected:
         PrimitiveType primitiveType_ = PrimitiveType::Triangles;
